chore: move testing fixtures to one single file
