+++ packages/execution/deps.ts
@@ -1,5 +1,5 @@
 export * from "../../deps.ts";
-export * from "../registry/mod.ts";
-export * from "../shared/mod.ts";
 export * from "../micro/proxy/mod.ts";
 export * from "../micro/router/mod.ts";
+export * from "../registry/mod.ts";
+export * from "../shared/mod.ts";
