--- packages/execution/src/testing/fixtures.ts
@@ -1,6 +1,0 @@
-import { appendCwd } from "../../deps.ts";
-
-export const writeScript = (scriptName: string, data: string) =>
-  Deno.writeTextFileSync(appendCwd(`/registry/${scriptName}.js`), data);
-export const removeScript = (scriptName: string) =>
-  Deno.removeSync(appendCwd(`/registry/${scriptName}.js`));
