+++ packages/shared/src/path_test.ts
@@ -1,5 +1,6 @@
+import { assertEquals } from "../../testing/mod.ts";
+
 import { appendCwd } from "./path.ts";
-import { assertEquals } from "./testing/deps.ts";
 
 Deno.test("should respect the forward slash", () => {
   assertEquals(appendCwd(`test`), appendCwd(`/test`));
