--- packages/shared/src/testing/deps.ts
@@ -1,1 +1,0 @@
-export * from "../../../testing/mod.ts";
