+++ packages/testing/mod.ts
@@ -1,1 +1,2 @@
 export * from "./deps.ts";
+export * from "./fixtures.ts";
